third_party/blink/web_tests/external/wpt/css/css-animations/KeyframeEffect-getKeyframes.tentative-expected.txt

This is a testharness.js-based test.
[FAIL] KeyframeEffect.getKeyframes() returns no frames for various kinds of empty animations
  assert_equals: number of frames when @keyframes has empty keyframes expected 0 but got 2
[FAIL] KeyframeEffect.getKeyframes() returns expected frames for an animation with multiple keyframes for the same time and with different easing functions and composite operations
  assert_array_equals: properties on Keyframe #1 should match lengths differ, expected array ["composite", "computedOffset", "easing", "marginTop", "offset", "paddingLeft"] length 6, got ["composite", "computedOffset", "easing", "fontSize", "offset"] length 5
[FAIL] KeyframeEffect.getKeyframes() returns expected values for animations with filter properties and missing keyframes
  assert_equals: value for 'filter' on Keyframe #1 should match expected "blur(5px) sepia(60%) saturate(30%)" but got "blur(5px) sepia(0.6) saturate(0.3)"
[FAIL] KeyframeEffect.getKeyframes() returns expected values for animations with background-size properties and missing keyframes
  assert_equals: value for 'backgroundSize' on ComputedKeyframe #1 should match expected "50% auto, 6px auto, contain" but got "50%, 6px, contain"
Harness: the test ran to completion.
